# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GalleryVault",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "summary": "Settings Get",
        "operationId": "settings_get_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Settings Get Api Settings Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Settings Save",
        "operationId": "settings_save_api_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Settings Save Api Settings Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/system/logs": {
      "get": {
        "summary": "System Logs Get",
        "description": "Retrieve recent system logs from the in-memory ring buffer.",
        "operationId": "system_logs_get_api_system_logs_get",
        "parameters": [
          {
            "name": "min_level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "INFO",
              "title": "Min Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response System Logs Get Api System Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "System Logs Clear",
        "description": "Clear the in-memory ring buffer logs.",
        "operationId": "system_logs_clear_api_system_logs_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response System Logs Clear Api System Logs Delete"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/logs/level": {
      "post": {
        "summary": "System Logs Set Level",
        "description": "Dynamically change the active log level without container restart.",
        "operationId": "system_logs_set_level_api_system_logs_level_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogLevelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response System Logs Set Level Api System Logs Level Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/saved-searches": {
      "get": {
        "summary": "Saved Searches List",
        "operationId": "saved_searches_list_api_saved_searches_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Saved Searches List Api Saved Searches Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Saved Searches Add",
        "operationId": "saved_searches_add_api_saved_searches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Saved Searches Add Api Saved Searches Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/saved-searches/{search_id}": {
      "delete": {
        "summary": "Saved Searches Delete",
        "operationId": "saved_searches_delete_api_saved_searches__search_id__delete",
        "parameters": [
          {
            "name": "search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Search Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Saved Searches Delete Api Saved Searches  Search Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogLevelRequest": {
        "properties": {
          "level": {
            "type": "string",
            "title": "Level",
            "description": "Log level: DEBUG, INFO, WARNING, ERROR, CRITICAL"
          }
        },
        "type": "object",
        "required": [
          "level"
        ],
        "title": "LogLevelRequest"
      },
...
      "SavedSearchRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "query": {
            "additionalProperties": true,
            "type": "object",
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SavedSearchRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}